# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ClinicFlow API Documentation",
        "description": "Comprehensive API documentation for ClinicFlow Backend",
        "contact": {
            "email": "admin@clinicflow.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        },
        {
            "url": "https://api.clinicflow.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get user notifications",
                "operationId": "b1274a2aaae8a4295a7e1925e8d8090f",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_read",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notifications list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/{id}/read": {
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark notification as read",
                "operationId": "8bc5203b95eaa5cf8614a66ced103386",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marked as read"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/mark-all-read": {
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark all notifications as read",
                "operationId": "48969f1c1523374341cf86b5610b3e03",
                "responses": {
                    "200": {
                        "description": "All marked as read"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/website/booking": {
            "post": {
                "tags": [
                    "Website"
                ],
                "summary": "Book appointment (public/guest)",
                "operationId": "a2b84f027def02242d54dd3e27900024",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "doctor_id",
                                    "clinic_id",
                                    "service_id",
                                    "date",
                                    "time"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "doctor_id": {
                                        "type": "integer"
                                    },
                                    "clinic_id": {
                                        "type": "integer"
                                    },
                                    "service_id": {
                                        "type": "integer"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "time": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created"
                    }
                }
            }
        },
        "/website/available-slots": {
            "get": {
                "tags": [
                    "Website"
                ],
                "summary": "Get available time slots",
                "operationId": "0e383a01b462bd4605690106f0ec23b1",
                "parameters": [
                    {
                        "name": "doctor_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Available slots"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}